backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
groq>=0.7.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.0.0
sentence-transformers>=3.0.0
huggingface-hub>=0.25.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
PyJWT>=2.8.0
slowapi>=0.1.9
redis>=5.0.0
sse-starlette>=1.8.0
python-multipart>=0.0.9
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
aiosqlite>=0.19.0
